tags component to rfc
